Quick and dirty connection to use ExtractImages.java instead of PdfImageExtractor.java for exporting images including density.
I found out the JPEG's in the PDF already contain the density X and Y, but lack the unit, which is 0. Those X's and Y's might be helpful in estimating the original density, and as the JFIF-chunk exists only modifying one byte for the unit in the stream could be a faster more lossless solution than a full ImageIOUtil.writeImage.